@@ -11,125 +12,140 @@
     "preview": "vite preview",
     "preview:build": "pnpm build && vite preview",
     "typecheck": "tsc --noEmit && vue-tsc --noEmit --skipLibCheck",
-    "svgo": "svgo -f src/assets/svg -o src/assets/svg",
+    "svgo": "svgo -f . -r",
     "cloc": "NODE_OPTIONS=--max-old-space-size=4096 cloc . --exclude-dir=node_modules --exclude-lang=YAML",
-    "clean:cache": "rimraf node_modules && rimraf .eslintcache && pnpm install",
+    "clean:cache": "rimraf .eslintcache && rimraf pnpm-lock.yaml && rimraf node_modules && pnpm store prune && pnpm install",
     "lint:eslint": "eslint --cache --max-warnings 0  \"{src,mock,build}/**/*.{vue,js,ts,tsx}\" --fix",
     "lint:prettier": "prettier --write  \"src/**/*.{js,ts,json,tsx,css,scss,vue,html,md}\"",
-    "lint:stylelint": "stylelint \"**/*.{html,vue,css,scss}\" --fix --cache --cache-location node_modules/.cache/stylelint/",
-    "lint:lint-staged": "lint-staged -c ./.husky/lintstagedrc.js",
-    "lint:pretty": "pretty-quick --staged",
+    "lint:stylelint": "stylelint --cache --fix \"**/*.{html,vue,css,scss}\" --cache-location node_modules/.cache/stylelint/",
     "lint": "pnpm lint:eslint && pnpm lint:prettier && pnpm lint:stylelint",
-    "prepare": "husky install",
+    "prepare": "husky",
     "preinstall": "npx only-allow pnpm"
   },
-  "browserslist": [
-    "> 1%",
-    "not ie 11",
-    "not op_mini all"
+  "keywords": [
+    "vue-pure-admin",
+    "element-plus",
+    "tailwindcss",
+    "pure-admin",
+    "typescript",
+    "pinia",
+    "vue3",
+    "vite",
+    "esm"
   ],
+  "homepage": "https://github.com/pure-admin/pure-admin-thin",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/pure-admin/pure-admin-thin.git"
+  },
+  "bugs": {
+    "url": "https://github.com/pure-admin/pure-admin-thin/issues"
+  },
+  "license": "MIT",
+  "author": {
+    "name": "xiaoxian521",
+    "email": "pureadmin@163.com",
+    "url": "https://github.com/xiaoxian521"
+  },
   "dependencies": {
-    "@pureadmin/descriptions": "^1.1.1",
-    "@pureadmin/table": "^2.3.2",
-    "@pureadmin/utils": "^1.9.6",
-    "@vueuse/core": "^10.2.0",
-    "@vueuse/motion": "^2.0.0",
+    "@pureadmin/descriptions": "^1.2.0",
+    "@pureadmin/table": "^3.1.2",
+    "@pureadmin/utils": "^2.4.5",
+    "@vueuse/core": "^10.9.0",
+    "@vueuse/motion": "^2.1.0",
     "animate.css": "^4.1.1",
-    "axios": "^1.4.0",
-    "dayjs": "^1.11.8",
-    "echarts": "^5.4.2",
-    "element-plus": "2.3.6",
+    "axios": "^1.6.7",
+    "dayjs": "^1.11.10",
+    "echarts": "^5.5.0",
+    "element-plus": "^2.6.0",
     "js-cookie": "^3.0.5",
-    "mitt": "^3.0.0",
-    "mockjs": "^1.1.0",
+    "localforage": "^1.10.0",
+    "mint-filter": "^4.0.3",
+    "mitt": "^3.0.1",
     "nprogress": "^0.2.0",
     "path": "^0.12.7",
-    "pinia": "^2.1.4",
-    "pinyin-pro": "^3.15.2",
+    "pinia": "^2.1.7",
+    "pinyin-pro": "^3.19.6",
     "qs": "^6.11.2",
     "responsive-storage": "^2.2.0",
-    "sortablejs": "^1.15.0",
-    "vue": "^3.3.4",
-    "vue-router": "^4.2.2",
-    "vue-types": "^5.1.0"
+    "sortablejs": "^1.15.2",
+    "vue": "^3.4.21",
+    "vue-i18n": "^9.10.1",
+    "vue-router": "^4.3.0",
+    "vue-tippy": "^6.4.1",
+    "vue-types": "^5.1.1"
   },
   "devDependencies": {
-    "@commitlint/cli": "^17.6.6",
-    "@commitlint/config-conventional": "^17.6.6",
+    "@commitlint/cli": "^18.6.1",
+    "@commitlint/config-conventional": "^18.6.2",
+    "@commitlint/types": "^18.6.1",
+    "@eslint/js": "^8.57.0",
+    "@faker-js/faker": "^8.4.1",
     "@iconify-icons/ep": "^1.2.12",
-    "@iconify-icons/ri": "^1.2.9",
+    "@iconify-icons/ri": "^1.2.10",
     "@iconify/vue": "^4.1.1",
-    "@pureadmin/theme": "^3.1.0",
-    "@types/js-cookie": "^3.0.3",
-    "@types/mockjs": "^1.0.7",
-    "@types/node": "^20.3.1",
-    "@types/nprogress": "0.2.0",
-    "@types/qs": "^6.9.7",
-    "@types/sortablejs": "^1.15.1",
-    "@typescript-eslint/eslint-plugin": "^5.60.0",
-    "@typescript-eslint/parser": "^5.60.0",
-    "@vitejs/plugin-vue": "^4.2.3",
-    "@vitejs/plugin-vue-jsx": "^3.0.1",
-    "@vue/eslint-config-prettier": "^7.1.0",
-    "@vue/eslint-config-typescript": "^11.0.3",
-    "autoprefixer": "^10.4.14",
+    "@intlify/unplugin-vue-i18n": "^2.0.0",
+    "@pureadmin/theme": "^3.2.0",
+    "@types/gradient-string": "^1.1.5",
+    "@types/js-cookie": "^3.0.6",
+    "@types/node": "^20.11.24",
+    "@types/nprogress": "^0.2.3",
+    "@types/qs": "^6.9.12",
+    "@types/sortablejs": "^1.15.8",
+    "@typescript-eslint/eslint-plugin": "^7.1.1",
+    "@typescript-eslint/parser": "^7.1.1",
+    "@vitejs/plugin-vue": "^5.0.4",
+    "@vitejs/plugin-vue-jsx": "^3.1.0",
+    "autoprefixer": "^10.4.18",
+    "boxen": "^7.1.1",
     "cloc": "^2.11.0",
-    "cssnano": "^6.0.1",
-    "eslint": "^8.43.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-vue": "^9.15.1",
-    "husky": "^8.0.3",
-    "lint-staged": "^13.2.2",
-    "picocolors": "^1.0.0",
-    "postcss": "^8.4.24",
-    "postcss-html": "^1.5.0",
-    "postcss-import": "^15.1.0",
-    "postcss-scss": "^4.0.6",
-    "prettier": "^2.8.8",
-    "pretty-quick": "^3.1.3",
-    "rimraf": "^5.0.1",
-    "rollup-plugin-visualizer": "^5.9.2",
-    "sass": "^1.63.6",
-    "sass-loader": "^13.3.2",
-    "stylelint": "^15.9.0",
-    "stylelint-config-html": "^1.1.0",
-    "stylelint-config-recess-order": "^4.2.0",
-    "stylelint-config-recommended": "^12.0.0",
-    "stylelint-config-recommended-scss": "^12.0.0",
-    "stylelint-config-recommended-vue": "^1.4.0",
-    "stylelint-config-standard": "^33.0.0",
-    "stylelint-config-standard-scss": "^9.0.0",
-    "stylelint-order": "^6.0.3",
-    "stylelint-prettier": "^3.0.0",
-    "stylelint-scss": "^5.0.1",
-    "svgo": "^3.0.2",
-    "tailwindcss": "^3.3.2",
-    "terser": "^5.18.1",
-    "typescript": "5.0.4",
-    "vite": "^4.3.9",
+    "cssnano": "^6.0.5",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-define-config": "^2.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-vue": "^9.22.0",
+    "gradient-string": "^2.0.2",
+    "husky": "^9.0.11",
+    "lint-staged": "^15.2.2",
+    "postcss": "^8.4.35",
+    "postcss-html": "^1.6.0",
+    "postcss-import": "^16.0.1",
+    "postcss-scss": "^4.0.9",
+    "prettier": "^3.2.5",
+    "rimraf": "^5.0.5",
+    "rollup-plugin-visualizer": "^5.12.0",
+    "sass": "^1.71.1",
+    "stylelint": "^16.2.1",
+    "stylelint-config-recess-order": "^5.0.0",
+    "stylelint-config-recommended-vue": "^1.5.0",
+    "stylelint-config-standard-scss": "^13.0.0",
+    "stylelint-prettier": "^5.0.0",
+    "svgo": "^3.2.0",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.3.3",
+    "vite": "^5.1.5",
     "vite-plugin-cdn-import": "^0.3.5",
     "vite-plugin-compression": "^0.5.1",
-    "vite-plugin-mock": "2.9.6",
-    "vite-plugin-remove-console": "^2.1.1",
-    "vite-svg-loader": "^4.0.0",
-    "vue-eslint-parser": "^9.3.1",
-    "vue-tsc": "^1.8.1"
+    "vite-plugin-fake-server": "^2.1.1",
+    "vite-plugin-remove-console": "^2.2.0",
+    "vite-plugin-router-warn": "^1.0.0",
+    "vite-svg-loader": "^5.1.0",
+    "vue-eslint-parser": "^9.4.2",
+    "vue-tsc": "^1.8.27"
   },
+  "engines": {
+    "node": "^18.18.0 || ^20.9.0 || >=21.1.0",
+    "pnpm": ">=8.6.10"
+  },
+  "packageManager": "pnpm@8.6.10",
   "pnpm": {
-    "peerDependencyRules": {
-      "ignoreMissing": [
-        "rollup",
-        "webpack",
-        "core-js"
-      ]
-    },
     "allowedDeprecatedVersions": {
       "sourcemap-codec": "*",
+      "domexception": "*",
       "w3c-hr-time": "*",
-      "stable": "*"
+      "stable": "*",
+      "abab": "*"
     }
-  },
-  "repository": "git@github.com:pure-admin/pure-admin-thin.git",
-  "author": "xiaoxian521",
-  "license": "MIT"
+  }
 }
