@@ -1,4 +1,7 @@
-module.exports = {
+// @ts-check
+
+/** @type {import('postcss-load-config').Config} */
+export default {
   plugins: {
     "postcss-import": {},
     "tailwindcss/nesting": {},
