@@ -1,5 +1,5 @@
 // 模拟后端动态生成路由
-import { MockMethod } from "vite-plugin-mock";
+import { defineFakeRoute } from "vite-plugin-fake-server/client";
 
 /**
  * roles：页面级别权限，这里模拟二种 "admin"、"common"
@@ -11,7 +11,7 @@ const permissionRouter = {
   path: "/permission",
   meta: {
     title: "权限管理",
-    icon: "lollipop",
+    icon: "ep:lollipop",
     rank: 10
   },
   children: [
@@ -29,15 +29,19 @@ const permissionRouter = {
       meta: {
         title: "按钮权限",
         roles: ["admin", "common"],
-        auths: ["btn_add", "btn_edit", "btn_delete"]
+        auths: [
+          "permission:btn:add",
+          "permission:btn:edit",
+          "permission:btn:delete"
+        ]
       }
     }
   ]
 };
 
-export default [
+export default defineFakeRoute([
   {
-    url: "/getAsyncRoutes",
+    url: "/get-async-routes",
     method: "get",
     response: () => {
       return {
@@ -46,4 +50,4 @@ export default [
       };
     }
   }
-] as MockMethod[];
+]);
