@@ -4,7 +4,7 @@ export interface ListItem {
   datetime: string;
   type: string;
   description: string;
-  status?: "" | "success" | "warning" | "info" | "danger";
+  status?: "primary" | "success" | "warning" | "info" | "danger";
   extra?: string;
 }
 
@@ -103,7 +103,7 @@ export const noticesData: TabItem[] = [
   },
   {
     key: "3",
-    name: "代办",
+    name: "待办",
     list: [
       {
         avatar: "",
