@@ -1,9 +1,9 @@
 <script setup lang="ts">
-import ArrowUpLine from "@iconify-icons/ri/arrow-up-line";
-import ArrowDownLine from "@iconify-icons/ri/arrow-down-line";
 import { useNav } from "@/layout/hooks/useNav";
 import mdiKeyboardEsc from "@/assets/svg/keyboard_esc.svg?component";
 import enterOutlined from "@/assets/svg/enter_outlined.svg?component";
+import ArrowUpLine from "@iconify-icons/ri/arrow-up-line";
+import ArrowDownLine from "@iconify-icons/ri/arrow-down-line";
 
 const props = withDefaults(defineProps<{ total: number }>(), {
   total: 0
@@ -50,7 +50,9 @@ const { device } = useNav();
     padding: 2px;
     margin-right: 3px;
     font-size: 20px;
-    box-shadow: inset 0 -2px #cdcde6, inset 0 0 1px 1px #fff,
+    box-shadow:
+      inset 0 -2px #cdcde6,
+      inset 0 0 1px 1px #fff,
       0 1px 2px 1px #1e235a66;
   }
 
