@@ -1,7 +1,6 @@
 <script setup lang="ts">
 import { SearchModal } from "./components";
 import { useBoolean } from "../../hooks/useBoolean";
-import Search from "@iconify-icons/ep/search";
 
 const { bool: show, toggle } = useBoolean();
 function handleSearch() {
@@ -10,11 +9,13 @@ function handleSearch() {
 </script>
 
 <template>
-  <div
-    class="search-container w-[40px] h-[48px] flex-c cursor-pointer navbar-bg-hover"
-    @click="handleSearch"
-  >
-    <IconifyIconOffline :icon="Search" />
+  <div>
+    <div
+      class="search-container w-[40px] h-[48px] flex-c cursor-pointer navbar-bg-hover"
+      @click="handleSearch"
+    >
+      <IconifyIconOffline icon="ri:search-line" />
+    </div>
+    <SearchModal v-model:value="show" />
   </div>
-  <SearchModal v-model:value="show" />
 </template>
