@@ -1,10 +1,11 @@
 import type { App } from "vue";
 import * as echarts from "echarts/core";
-import { CanvasRenderer } from "echarts/renderers";
 import { PieChart, BarChart, LineChart } from "echarts/charts";
+import { CanvasRenderer, SVGRenderer } from "echarts/renderers";
 import {
   GridComponent,
   TitleComponent,
+  PolarComponent,
   LegendComponent,
   GraphicComponent,
   ToolboxComponent,
@@ -20,8 +21,10 @@ use([
   BarChart,
   LineChart,
   CanvasRenderer,
+  SVGRenderer,
   GridComponent,
   TitleComponent,
+  PolarComponent,
   LegendComponent,
   GraphicComponent,
   ToolboxComponent,
@@ -31,9 +34,8 @@ use([
 ]);
 
 /**
- * @description 按需引入echarts
- * @see {@link https://echarts.apache.org/handbook/zh/basics/import#%E6%8C%89%E9%9C%80%E5%BC%95%E5%85%A5-echarts-%E5%9B%BE%E8%A1%A8%E5%92%8C%E7%BB%84%E4%BB%B6}
- * @see 温馨提示：必须将 `$echarts` 添加到全局 `globalProperties` ，为了配合 https://pure-admin-utils.netlify.app/hooks/useEcharts/useEcharts.html 使用
+ * @description 按需引入echarts，具体看 https://echarts.apache.org/handbook/zh/basics/import/#%E5%9C%A8-typescript-%E4%B8%AD%E6%8C%89%E9%9C%80%E5%BC%95%E5%85%A5
+ * @see 温馨提示：必须将 `$echarts` 添加到全局 `globalProperties` ，具体看 https://pure-admin-utils.netlify.app/hooks/useECharts/useECharts#%E4%BD%BF%E7%94%A8%E5%89%8D%E6%8F%90
  */
 export function useEcharts(app: App) {
   app.config.globalProperties.$echarts = echarts;
