@@ -2,9 +2,9 @@ export default {
   path: "/error",
   redirect: "/error/403",
   meta: {
-    icon: "informationLine",
-    title: "异常页面",
+    icon: "ri:information-line",
     // showLink: false,
+    title: "异常页面",
     rank: 9
   },
   children: [
@@ -33,4 +33,4 @@ export default {
       }
     }
   ]
-} as RouteConfigsTable;
+} satisfies RouteConfigsTable;
