@@ -27,4 +27,4 @@ export default [
       }
     ]
   }
-] as Array<RouteConfigsTable>;
+] satisfies Array<RouteConfigsTable>;
