@@ -1,4 +1,4 @@
-import { RouteRecordName } from "vue-router";
+import type { RouteRecordName } from "vue-router";
 
 export type cacheType = {
   mode: string;
@@ -38,4 +38,6 @@ export type setType = {
 export type userType = {
   username?: string;
   roles?: Array<string>;
+  isRemembered?: boolean;
+  loginDay?: number;
 };
