@@ -1,5 +1,5 @@
 declare module "*.vue" {
-  import { DefineComponent } from "vue";
+  import type { DefineComponent } from "vue";
   const component: DefineComponent<{}, {}, any>;
   export default component;
 }
@@ -8,3 +8,5 @@ declare module "*.scss" {
   const scss: Record<string, string>;
   export default scss;
 }
+
+declare module "element-plus/dist/locale/zh-cn.mjs";
